fix(text-extraction): satisfy architecture conventions
Three architecture-test violations surfaced on CI:

- Exception_classes_should_reside_in_Exceptions_folder — move
  TextExtractionException.cs into Exceptions/ (namespace
  Granit.TextExtraction.Exceptions).
- Options_classes_should_not_be_at_module_root — move
  GranitTextExtractionOptions.cs into Options/ (namespace
  Granit.TextExtraction.Options).
- SectionName_must_be_hierarchical — TextExtraction is the top-level
  module name, so add it to AllowedSingleSegmentSections alongside
  BackgroundJobs, QueryEngine, BlobStorage, etc.

Consumer files updated to import the new sub-namespaces; test files use
type aliases to keep call sites readable in the presence of the new
Granit.TextExtraction.Options namespace.
