+++ tests/Granit.ArchitectureTests/SectionNameConventionTests.cs
@@ -47,6 +47,7 @@ public sealed partial class SectionNameConventionTests
         "QueryEngine",
         "RateLimiting",
         "Settings",
+        "TextExtraction",
         "Vault",
         "Webhooks",
         "Wolverine",
