+++ tests/Granit.TextExtraction.Tests/GranitTextExtractionOptionsTests.cs
@@ -1,3 +1,4 @@
+using Granit.TextExtraction.Options;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
