+++ tests/Granit.TextExtraction.Tests/LimitedStreamTests.cs
@@ -1,4 +1,5 @@
 using System.Text;
+using Granit.TextExtraction.Exceptions;
 using Shouldly;
 using Xunit;
